\name{Quadra}
\alias{Quadra}
\title{Object of class Quadra}
\description{
Create an object of class Quadra
}
\usage{
Quadra(a,b,c)
}
\arguments{
 \item{a}{Coefficient A (see details)}
 \item{b}{Coefficient B (see details)}
 \item{c}{Coefficient C (see details)}
}
\details{
The polynomial is \eqn{f(x)=Ax^2+Bx+C}.
}

\examples{
P1 <- Quadra(1,2,3)
P1
}

