\name{summary}
\alias{summary.Quadra}
\title{Summary for objects of class Quadra}
\description{
Returns characteristics of the polynomial
}
\usage{
\method{summary}{Quadra}(object, ...)
}
\arguments{
 \item{object}{An object of class \code{Quadra}}
\item{...}{Argument for other type of objects}
}

\examples{
P1 <- Quadra(1,2,3)
summary(P1)

}

